feat: save user in local storage
